+++ Photo Club Hub/ViewModel/Organization.swift
@@ -28,12 +28,17 @@ extension Organization {
 	}
 
     var organizationType: OrganizationType {
+        @MainActor
         get { // careful: cannot read organizationType on background thread if database still contains nil
             if organizationType_ != nil {
                 return organizationType_! // organizationType_ cannot be nil at this point
             } else {
                 // something is fundamentally wrong if this happens
-                fatalError( "Error because organization is nil", file: #file, line: #line )
+                ifDebugFatalError( "Error because organization is nil", file: #file, line: #line )
+                let persistenceController = PersistenceController.shared // for Core Data
+                let viewContext = persistenceController.container.viewContext // requires @MainActor
+                // swiftlint:disable:next force_cast
+                return viewContext.registeredObject(for: OrganizationType.enum2objectID[.unknown]!) as! OrganizationType
             }
         }